Split SLSA v02 pred to own file
